Refactor custom views out of urls.py
